% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualPlotFunctions.R
\name{add_sample_size_labels}
\alias{add_sample_size_labels}
\title{Add Sample Size Information to Axis Labels}
\usage{
add_sample_size_labels(data, x_var)
}
\arguments{
\item{data}{A data frame}

\item{x_var}{Character string. Column name for the categorical variable}
}
\value{
A data frame with updated labels
}
\description{
Adds sample size (N) information to categorical axis labels.
}
