% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductionPlots.r
\name{plot_tsne}
\alias{plot_tsne}
\title{Create a T-distributed stochastic neighbor embedding plot}
\usage{
plot_tsne(dataframe, title = "", feature)
}
\arguments{
\item{title}{Character string. Plot title (default: "")}

\item{feature}{Named array. Class labels of the samples that will be
shown under different colors in the plot.}

\item{data}{A numeric data frame or matrix. Rows are samples, columns are
features}
}
\value{
A ggplot2 object
}
\description{
Creates a T-distributed stochastic neighbor embedding plot using
the M3C library tsne function
}
\examples{
\dontrun{
plot_pca(data = expression_matrix,
         feature_list = list(group = sample_groups, batch = sample_batches))
}
}
